Input widget have his own label
This is probably architecturally incorrect, and it would be more correct
to have separate "label" and "input" elements as in HTML. However, when
creating forms, almost all input widgets assume the optional presence of
a label. Instead of copying the code that adds the label each time, we
will add the label directly to the input widget.

Signed-off-by: Alexey Gladkov <legion@kernel.org>