Fix the output of the wchar strings
ncurses provides a set of functions for writing formatted output to
a curses window. But the problem is that there are no versions of these
functions for wide characters.

The man page note says:

  To format and write a wide-character string to a curses window,
  consider using swprintf(3) and waddwstr(3X) or similar.

Another problem is that swprintf (at least in glibc) does not support
the behavior of `snprintf(NULL, 0, ...)`.

Therefore, it will be necessary to create your own wrappers to output
such data.

Signed-off-by: Alexey Gladkov <legion@kernel.org>